feat(loadPattern): Rename pattern_assembler to loadPattern and simplify export
