+++ core/lib/processRecursive.js
@@ -23,7 +23,7 @@ module.exports = function (file, patternlab) {
   //call our helper method to actually unravel the pattern with any partials
   return decompose(currentPattern, patternlab)
     .catch(reason => {
-      console.log(reason)
+      console.log(reason);
       logger.error(reason);
     });
 };
